Remove deprecated MAINTAINER and add STOPSIGNAL

+++ bin/start_consul.sh
@@ -3,8 +3,13 @@
 log() {
 	printf '%s\n' "$@"|awk '{print strftime("%FT%T%z",systime()),"[INFO] start_consul.sh:",$0}'
 }
+
 loge() {
 	printf '%s\n' "$@"|awk '{print strftime("%FT%T%z",systime()),"[ERROR] start_consul.sh:",$0}' >&2
+}
+
+set_consul_performance() {
+	PERFORMANCE=$1 su-exec consul sh -c "{ rm /etc/consul/consul.json;jq '.performance.raft_multiplier=(env.PERFORMANCE|tonumber)' >/etc/consul/consul.json; } < /etc/consul/consul.json"
 }
 
 # Add Consul FQDN to hosts file for convenience
@@ -12,35 +17,35 @@ printf '%s\t%s\n' "$(hostname -i)" "$(hostname).node.${CONSUL_DNS:-consul}" >> /
 
 # Performance configuration for Cloud providers
 # See https://www.consul.io/docs/guides/performance.html
-if [ "${CONSUL_ENVIRONMENT:-non-prod}" = 'prod' ]; then
-	su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 1' > /etc/consul/consul.json; } < /etc/consul/consul.json"
-else
+if [ "${CONSUL_ENVIRONMENT:-auto}" = 'max' ]; then
+	set_consul_performance 1
+elif [ "${CONSUL_ENVIRONMENT:-auto}" = 'auto' ]; then
 	# Amazon EC2
 	if [ -f /sys/hypervisor/uuid ] && [ "$(head -c 3 /sys/hypervisor/uuid | tr '[:lower:]' '[:upper:]')" = 'EC2' ]; then
 		EC2_INSTANCE-TYPE="$(wget -q -O- 'http://169.254.169.254/latest/meta-data/instance-type' | awk -F. '{print $2}')"
-		case "$EC2_INSTANCE-TYPE" in
-			nano)   su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 6' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			micro)  su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 5' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			small)  su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 3' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			medium) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 2' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			*large) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 1' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
+		case "${EC2_INSTANCE-TYPE}" in
+			nano)   set_consul_performance 6 ;;
+			micro)  set_consul_performance 5 ;;
+			small)  set_consul_performance 3 ;;
+			medium) set_consul_performance 2 ;;
+			*large) set_consul_performance 1 ;;
 		esac
 	# GCE
 	elif [ -f /sys/class/dmi/id/bios_vendor ] && grep -iq 'Google' /sys/class/dmi/id/bios_vendor; then
-		GCE_MACHINE-TYPE="$(wget -q -O- 'http://metadata.google.internal/computeMetadata/v1/instance/machine-type')"
-		case "$GCE_MACHINE-TYPE" in
-			f1-micro) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 5' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			g1-small) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 3' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			*)        su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 1' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
+		GCE_MACHINE-TYPE="$(wget -q -O- --header 'Metadata-Flavor: Google' 'http://metadata.google.internal/computeMetadata/v1/instance/machine-type'| awk -F/ '{print $NF}')"
+		case "${GCE_MACHINE-TYPE}" in
+			f1-micro) set_consul_performance 5 ;;
+			g1-small) set_consul_performance 3 ;;
+			*)        set_consul_performance 1 ;;
 		esac
 	# MS Azure
-	elif grep -iq 'Hyper-V UEFI' /sys/class/dmi/id/bios_version; then
+	elif [ -f /sys/class/dmi/id/bios_vendor ] && grep -iq 'Hyper-V UEFI' /sys/class/dmi/id/bios_version; then
 		AZURE_VMSIZE="$(wget -q -O- --header 'Metadata: true' 'http://169.254.169.254/metadata/instance/compute/vmSize?api-version=2017-04-02')"
-		case "$AZURE_VMSIZE" in
-			Standard_A0) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 5' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			Standard_A1) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 3' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			Standard_A2) su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 2' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
-			*)           su -s/bin/sh consul -c "{ rm /etc/consul/consul.json; jq '.performance.raft_multiplier = 1' > /etc/consul/consul.json; } < /etc/consul/consul.json" ;;
+		case "${AZURE_VMSIZE}" in
+			Standard_A0) set_consul_performance 5 ;;
+			Standard_A1) set_consul_performance 3 ;;
+			Standard_A2) set_consul_performance 2 ;;
+			*)           set_consul_performance 1 ;;
 		esac
 	else
 		loge "Can't determine performance settings. Using Consul default settings"
@@ -66,16 +71,15 @@ if [ "$CONSUL_LOWEST_PORT" -le 1024 ]; then
 	fi
 fi
 
-if [ -e /data/raft/raft.db ]; then
+if [ -e /data/raft/raft.db ]; then # This is a restart
 	consul validate /etc/consul/consul.json || exit 1
-	# This is a restart
 	log 'Starting Consul'
 	unset CONSUL_ENCRYPT_TOKEN
 	unset CONSUL_BOOTSTRAP_HOST
 	unset CONSUL_CLUSTER_SIZE
 
 	exec su-exec consul:consul consul agent -server -ui -config-dir=/etc/consul/ -datacenter="$CONSUL_DC_NAME" -domain="${CONSUL_DOMAIN:-consul}" -retry-join="$CONSUL_DNS_NAME" -rejoin
-else
+else # This is the first start
 	if [ "$CONSUL_DC_NAME" ] && [ "$CONSUL_ENCRYPT_TOKEN" ] && [ "$CONSUL_CLUSTER_SIZE" ] && [ "$CONSUL_DNS_NAME" ]; then
 		log 'Starting Consul for the first time, using CONSUL_DC_NAME & BOOTSTRAP_HOST environment variables'
